==> 1_Trim_Reivew.ipynb <==
# 論文資料及整理
* YELP
    * EDA
    * 50 review above 
    * join business, check-in and tips

# Import

import pandas as pd
import json
import numpy as np
from collections import Counter
import numpy as np
import pickle

res_df = pd.read_pickle ('../data/ORI_by_postal_code_s_dropped.pkl')
postal_code_feature_dict = pickle.load(open("../data/postal_dict.pkl", "rb"))
all_df = pd.read_pickle('../Data/restaurant_only.pkl')

len(Counter(res_df.business_id))

smaller_reviews = pd.DataFrame()
total_cnt = 0
for id in Counter(res_df.business_id):
    if total_cnt % 100 == 0:
        print(f'Now progress ... {total_cnt}')
    
    total_cnt +=1

    tmp = all_df[all_df.business_id == id ].sort_values(by='date',ascending=False)
    if len(tmp) >= 30:
        smaller_reviews = pd.concat([smaller_reviews , tmp[:30]])
    else:
        smaller_reviews = pd.concat([smaller_reviews , tmp])



smaller_reviews[:10]

smaller_reviews.to_pickle('../Data/restaurant_only_s.pkl')
