==> 3_Topic_Modeling.ipynb <==
# Create the features from Topic Modeling
* using by_postal_code dataset

# Import

import pandas as pd
import json
import numpy as np
from collections import Counter
import matplotlib.pyplot as plt
import pickle
import math 
from sklearn.model_selection import train_test_split

from scipy import sparse as sp
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import LabelEncoder
from sklearn.feature_selection import chi2
from sklearn.feature_selection import SelectKBest

RANDOM_STATE = 24

# 導入資料
train_df = pd.read_pickle('/home/adam/Steph_C/my_thesis/data/Train_by_postoal_code_without_review_pointwise_v3_3.pkl')
test_df = pd.read_pickle('/home/adam/Steph_C/my_thesis/data/Test_by_postoal_code_without_review_pointwise_v3_3.pkl')
all_df = pd.read_pickle('../Data/yelp/restaurant_only.pkl')

train_df = train_df.reset_index(drop=True)
test_df = test_df.reset_index(drop=True)
all_df = all_df.reset_index(drop=True)

print(train_df.shape , test_df.shape , all_df.shape)

# Create Tokens

import spacy
import pyLDAvis.gensim_models
from gensim.corpora.dictionary import Dictionary
from gensim.models import LdaMulticore
from gensim.models import CoherenceModel
from gensim import matutils

from nltk.stem import PorterStemmer
import nltk
from nltk.corpus import stopwords

# # EXAMPLE 
# import spacy
# from spacy.lang.en.examples import sentences 

# # nlp = spacy.load("en_core_web_md")
# # doc = nlp(sentences[0])
# # print(doc.text)
# # for token in doc:
# #     print(token.text, token.pos_, token.dep_)

# print(sentences)

# dictionary = Dictionary([preprocess(''.join(sentences))])
# print(dictionary)
# # dictionary.filter_extremes(no_below=5, no_above=0.9, keep_n=100)
# corpus = [dictionary.doc2bow(doc) for doc in [preprocess(''.join(sentences))]]
# print(corpus[0])


nlp = spacy.load("en_core_web_md")
keep_pos = ['NOUN']
# removal= ['ADV','PRON','CCONJ','PUNCT','PART','DET','ADP','SPACE', 'NUM', 'SYM']

def preprocess(text : str) -> list:
    
    for summary in nlp.pipe([text]):
        proj_tok = [token.lemma_.lower() for token in summary \
                    if token.pos_ in keep_pos and not token.is_stop and token.is_alpha]

    return proj_tok

def get_attribute_tokens(df):
    
    attribute_list = ['Ambience' , 'GoodForMeal' ]
    att_list = []
    df = df.reset_index(drop= True)
    
    for i in range(len(df)):
        try:
            for k,v in df.attributes[i].items():
                if k in attribute_list:
                    if isinstance(v,dict):
                        for k_1 , v_1 in v.items():
                            if v_1:
                                att_list.append(k_1)
        except:
            pass

    return att_list

# Create tokens
def get_tokens(df):
    
    df['res_tokens'] = ''
    df['loc_tokens'] = ''
    df['res_tokens_with_att'] = ''
    df['loc_tokens_with_att'] = ''
    
    for i in range(len(df)):
        
        if i%1000==0:
            print(f'Now progress .... {i}')
        
        res_str = ''
        loc_str = ''

        res = df[df.name==df.name[i]]
        res = res[res.postal_code != df.postal_code[i]]

        loc = df[df.postal_code==df.postal_code[i]]
        loc = loc[loc.name != df.name[i]]

        # get res_string
        for j in Counter(res.text):
            removed = []
            if j not in removed:
                res_str+=j+' '
                removed.append(j)

        # get loc_string
        for j in Counter(loc.text):
            removed = []
            if j not in removed:
                loc_str+=j+' '
                removed.append(j)
        
        # get attributes tokens
        res_att = get_attribute_tokens(res)
        loc_att = get_attribute_tokens(loc)
        
        orig_list_res = preprocess(res_str)
        orig_list_loc = preprocess(loc_str)
        
        df['res_tokens'][i] = orig_list_res
        df['loc_tokens'][i] = orig_list_loc
        
        df['res_tokens_with_att'][i] = orig_list_res+res_att
        df['loc_tokens_with_att'][i] = orig_list_loc+loc_att
        
        
        
    
    print(f'Finish building .....')
    

    return df
    

# Takes a LONG time
train_df = get_tokens(train_df)
test_df = get_tokens(test_df)

# LDA

def get_LDA(train_df , test_df ,tokens):
    
    train_df['LDA_'+tokens] = ''
    test_df['LDA_'+tokens] = ''
    dictionary = Dictionary(train_df[tokens])
    dictionary.filter_extremes(no_below=5, no_above=0.5, keep_n=1000)
    
    #### doc2Bow
    train_corpus = [dictionary.doc2bow(doc) for doc in train_df[tokens]]
    test_corpus = [dictionary.doc2bow(doc) for doc in test_df[tokens]]

    # building models
    lda_model = LdaMulticore(corpus=train_corpus, id2word=dictionary, iterations=50, \
                             num_topics=10, workers = 4, passes=10)
    
    for i in range(len(train_corpus)):
        
        if i%1000==0:
            print(f'Now progress .... {i}')
        
        train_df['LDA_'+tokens][i] = lda_model[train_corpus][i]

    print(f'Finish building train .....')
    
    for i in range(len(test_corpus)):
        
        if i%1000==0:
            print(f'Now progress .... {i}')
        
        test_df['LDA_'+tokens][i] = lda_model[test_corpus][i]

    print(f'Finish building test .....')
    return train_df , test_df , lda_model

def get_LDA_tfidf(train_df , test_df,tokens ):
    
    train_df['LDA_'+tokens+'_tfidf'] = ''
    test_df['LDA_'+tokens+'_tfidf'] = ''

    vectorizer = TfidfVectorizer(min_df=10, max_df=0.2, max_features=1000)
    
    #### TF-IDF
    train_documents = [' '.join(tokens) for tokens in list(train_df[tokens])]
    test_documents = [' '.join(tokens) for tokens in list(test_df[tokens])]

    train_corpus = matutils.Sparse2Corpus(vectorizer.fit_transform(train_documents).T)
    test_corpus = matutils.Sparse2Corpus(vectorizer.fit_transform(test_documents).T)
    
    
    # building models
    id2word = dict((v, k) for k, v in vectorizer.vocabulary_.items())
    lda_model = LdaMulticore(corpus=train_corpus, id2word=id2word, iterations=50, \
                             num_topics=10, workers = 4, passes=10)
    
    for i in range(len(train_corpus)):
        
        if i%1000==0:
            print(f'Now progress .... {i}')
        
        train_df['LDA_'+tokens+'_tfidf'][i] = lda_model[train_corpus][i]

    print(f'Finish building train .....')
    
    for i in range(len(test_corpus)):
        
        if i%1000==0:
            print(f'Now progress .... {i}')
        
        test_df['LDA_'+tokens+'_tfidf'][i] = lda_model[test_corpus][i]

    print(f'Finish building test .....')
    return train_df , test_df , lda_model

train_df ,test_df , res_lda = get_LDA(train_df ,test_df , 'res_tokens')
train_df ,test_df , res_lda_w_attr = get_LDA(train_df ,test_df , 'res_tokens_with_att')
train_df ,test_df , loc_lda = get_LDA(train_df ,test_df , 'loc_tokens')
train_df ,test_df , loc_lda_w_attr = get_LDA(train_df ,test_df , 'loc_tokens_with_att')

train_df ,test_df , res_lda_tfidf = get_LDA_tfidf(train_df ,test_df , 'res_tokens_with_att')
train_df ,test_df , loc_lda_tfidf = get_LDA_tfidf(train_df ,test_df , 'loc_tokens_with_att')

res_lda_tfidf.print_topics(-1)

loc_lda_tfidf.print_topics(-1)

## Cal Cosine Similarity 

# from gensim.matutils import cossim
# # Return cosine similarity between two sparse vectors. 
# # The similarity is a number between <-1.0, 1.0>, higher is more similar.
# print(cossim(train_df.LDA_res_tokens[1], train_df.LDA_loc_tokens[1]))

train_df.columns

# create Affinity and Complementary
from gensim.matutils import cossim

def get_cos_sim(df):
    
    df['tok_cosine_sim'] =''
    df['tok_cosine_sim_w_attr'] =''
    df['tok_cosine_sim_tfidf'] =''
    
    for i in range(len(df)):
        df['tok_cosine_sim'][i] = cossim(df.LDA_res_tokens[i], df.LDA_loc_tokens[i])
        df['tok_cosine_sim_w_attr'][i] = cossim(df.LDA_res_tokens_with_att[i], df.LDA_loc_tokens_with_att[i])
        df['tok_cosine_sim_tfidf'][i] = cossim(df.LDA_res_tokens_with_att_tfidf[i], df.LDA_loc_tokens_with_att_tfidf[i])
    
    return df

train_df = get_cos_sim(train_df)
test_df = get_cos_sim(test_df)

## Affinity & Complementary

def get_aff_comp(df) : 
    
    df['tok_affinity'] = ''
    df['tok_complementary'] = ''
    df['tok_affinity_with_att'] = ''
    df['tok_complementary_with_att'] = ''
    df['tok_affinity_tfidf'] = ''
    df['tok_complementary_tfidf'] = ''
    
    for i in range(len(df)):
        affinity = 0
        complementary = 0
        affinity_with_att = 0
        complementary_with_att = 0
        affinity_tfidf = 0
        complementary_tfidf = 0
        
        for j in df.LDA_res_tokens[i]:
            for k in df.LDA_loc_tokens[i]: 
                if j[0] in k:
                    affinity += j[1]*k[1]
                    complementary += j[1]*(1-k[1])
        
        
        for j in df.LDA_res_tokens_with_att[i]:
            for k in df.LDA_loc_tokens_with_att[i]: 
                if j[0] in k:
                    affinity_with_att += j[1]*k[1]
                    complementary_with_att += j[1]*(1-k[1]) 
        
        for j in df.LDA_res_tokens_with_att_tfidf[i]:
            for k in df.LDA_loc_tokens_with_att_tfidf[i]: 
                if j[0] in k:
                    affinity_tfidf += j[1]*k[1]
                    complementary_tfidf += j[1]*(1-k[1]) 
        
        df['tok_affinity'][i] = affinity
        df['tok_complementary'][i] = complementary
        df['tok_affinity_with_att'][i] = affinity_with_att
        df['tok_complementary_with_att'][i] = complementary_with_att
        df['tok_affinity_tfidf'][i] = affinity_tfidf
        df['tok_complementary_tfidf'][i] = complementary_tfidf
        
    return df

        

train_df = get_aff_comp(train_df)
test_df = get_aff_comp(test_df)

for i in ['tok_affinity','tok_complementary','tok_affinity_with_att','tok_complementary_with_att','tok_cosine_sim','tok_affinity_tfidf','tok_complementary_tfidf']:
    train_df[i] = train_df[i].astype('float')
    test_df[i] = test_df[i].astype('float')

train_df.to_pickle('/home/adam/Steph_C/my_thesis/data/Train_by_postoal_code_pointwise_v3_3_tfidf.pkl')
test_df.to_pickle('/home/adam/Steph_C/my_thesis/data/Test_by_postoal_code_pointwise_v3_3_tfidf.pkl')
