==> 4_Model_Linear_model.ipynb <==
# Linear Regression Benchmark
* SVR
* [Ridge](http://www.science.smith.edu/~jcrouser/SDS293/labs/lab10-py.html)
* GRBT

# Import

import numpy as np  
import pandas as pd 
from sklearn.model_selection import train_test_split
import matplotlib.pyplot as plt 
from sklearn.preprocessing import scale 
from sklearn import preprocessing
from sklearn.preprocessing import StandardScaler 
from sklearn.metrics import r2_score, mean_absolute_error, mean_squared_error

# models
from sklearn.metrics import mean_squared_error


# 導入資料
drill = pd.read_csv(data_path) #drill為原始資料檔
df=pd.DataFrame(drill)

# Functions

# create dataset

#模型評估
print('\nThe value of default measurement of rbf SVR is', svr_rbf.score(x_test, y_test))
print('R-squared value of rbf SVR is', r2_score(y_test, svr_predict))
print('The mean squared error of rbf SVR is', mean_squared_error(ss_y.inverse_transform(y_test), ss_y.inverse_transform(svr_predict)))
print('The mean absolute error of rbf SVR is', mean_absolute_error(ss_y.inverse_transform(y_test), ss_y.inverse_transform(svr_predict)))
print(svr_rbf.intercept_) #迴歸模型的截距

#輸出圖
plt.scatter(svr_predict, y_test, color= 'black', label= 'Data') #數據點
plt.plot(svr_predict, svr_predict, color= 'red', label= 'RBF model') #迴歸線
plt.xlabel('data')
plt.ylabel('quality')
plt.title('Support Vector Regression')
plt.legend()
plt.show()

# 多組比較圖
#set up plotting area
plt.figure(0).clf()

#fit logistic regression model and plot ROC curve
model = LogisticRegression()
model.fit(X_train, y_train)
y_pred = model.predict_proba(X_test)[:, 1]
fpr, tpr, _ = metrics.roc_curve(y_test, y_pred)
auc = round(metrics.roc_auc_score(y_test, y_pred), 4)
plt.plot(fpr,tpr,label="Logistic Regression, AUC="+str(auc))

#fit gradient boosted model and plot ROC curve
model = GradientBoostingClassifier()
model.fit(X_train, y_train)
y_pred = model.predict_proba(X_test)[:, 1]
fpr, tpr, _ = metrics.roc_curve(y_test, y_pred)
auc = round(metrics.roc_auc_score(y_test, y_pred), 4)
plt.plot(fpr,tpr,label="Gradient Boosting, AUC="+str(auc))

#add legend
plt.legend()


# mse = mean_squared_error(ytest,ypred)
# print(“MSE: %.2f” % mse)

# x_ax = range(len(ytest))
# plt.scatter(x_ax, ytest, s=5, color=”blue”, label=”original”)
# plt.plot(x_ax, ypred, lw=0.8, color=”red”, label=”predicted”)
# plt.legend()
# plt.show()

# SVR

from sklearn.svm import SVR  
x=df.loc[:,['Spindle speed', 'Feed rate', 'Point angle']] 
y=df['Thrust force']

# 資料分割
x_train, x_test, y_train, y_test = train_test_split(x.values, y.values, random_state=10, train_size=0.8)

#正規化
ss_x = StandardScaler()
ss_y = StandardScaler()
x_train = ss_x.fit_transform(x_train)
x_test = ss_x.fit_transform(x_test)
y_train = ss_y.fit_transform(y_train.reshape(-1, 1))
y_test = ss_y.fit_transform(y_test.reshape(-1, 1))

#建立SVR
svr_rbf = SVR(C=1e3, kernel='rbf', gamma='auto')
svr_rbf.fit(x_train,y_train)
svr_predict=svr_rbf.predict(x_test)


# Ridge

from sklearn.linear_model import Ridge, RidgeCV, Lasso, LassoCV
X, y = datasets.load_diabetes(return_X_y=True)
X_train, X_test, y_train, y_test = model_selection.train_test_split(X, y)

model = Ridge()
model.fit(X_train, y_train)
y_pred = model.predict(X_test)

# GBRT

from sklearn.ensemble import GradientBoostingRegressor

xtrain, xtest, ytrain, ytest=train_test_split(x, y, random_state=12,
test_size=0.15)
# with new parameters
gbr = GradientBoostingRegressor(n_estimators=600,
max_depth=5,
learning_rate=0.01,
min_samples_split=3)
# with default parameters
gbr = GradientBoostingRegressor()

gbr.fit(xtrain, ytrain)

ypred = gbr.predict(xtest)
